+++ metatensor-core/CHANGELOG.md
@@ -23,6 +23,12 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 - `mts_labels_difference` for finding the set difference between two `mts_labels_t`
 
+### metatensor-core C++
+
+#### Added
+
+- `Labels::set_difference` for finding the set difference between two `Labels`
+
 ## [Version 0.1.12](https://github.com/metatensor/metatensor/releases/tag/metatensor-core-v0.1.12) - 2025-02-17
 
 ### Changed